+++ src/main/java/hng_java_boilerplate/plans/dtos/PlanResponse.java
@@ -1,6 +1,6 @@
 package hng_java_boilerplate.plans.dtos;
 
-import hng_java_boilerplate.plans.Plan;
+import hng_java_boilerplate.plans.entity.Plan;
 
 public record PlanResponse(
         Plan data,
