+++ src/main/java/hng_java_boilerplate/plans/repository/PlanRepository.java
@@ -1,5 +1,6 @@
-package hng_java_boilerplate.plans;
+package hng_java_boilerplate.plans.repository;
 
+import hng_java_boilerplate.plans.entity.Plan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
