+++ src/test/java/hng_java_boilerplate/plan_test/PlanServiceTest.java
@@ -1,8 +1,8 @@
 package hng_java_boilerplate.plan_test;
 
-import hng_java_boilerplate.plans.Plan;
-import hng_java_boilerplate.plans.PlanRepository;
-import hng_java_boilerplate.plans.PlanServiceImpl;
+import hng_java_boilerplate.plans.entity.Plan;
+import hng_java_boilerplate.plans.repository.PlanRepository;
+import hng_java_boilerplate.plans.serviceImpl.PlanServiceImpl;
 import hng_java_boilerplate.plans.dtos.CreatePlanDto;
 import hng_java_boilerplate.plans.dtos.PlanResponse;
 import hng_java_boilerplate.plans.exceptions.DuplicatePlanException;
